Fixed bug where junction doesn't work after reset
The bug was caused by the set state function not being overriden, and thus, the u/d state (implemented by Junction) was not being reset to 0. The updates weren't propogated past the junction since updates are only propogated if states are changed, which it wasn't since it was hardstuck to 0.